revert back integer length to 20; remove usertypes, instead use roles

+++ database/migrations/2020_06_26_024404_create_system_configs_table.php
@@ -16,8 +16,8 @@ class CreateSystemConfigsTable extends Migration
         Schema::create('system_configs', function (Blueprint $table) {
             $table->string('keyword');
             $table->text('keyvalue')->nullable();
-            $table->unsignedInteger('created_by');
-            $table->unsignedInteger('updated_by')->nullable();
+            $table->unsignedBigInteger('created_by');
+            $table->unsignedBigInteger('updated_by')->nullable();
             $table->timestamps();
             
             $table->primary(['keyword']);
